Update index.html
Added functionalities from below requirements:Now don't make any changes to the current script but make sure the Mute Myself functionality is working. And I don't think the Approach for Chant Synchronization is working.  Make sure to fix that. And tell me how to test how it is working. Finally, the submitted comments should be visible to every one with a time stamp of when it was submitted.
In all of this make sure you keep the all the current functionality working and don't mess with that.