Update script.js
Below are the requirements added:
I still don't understand what the current mute and unmute button functionality does. 
Now don't make any changes to the current script and preserve all previous functionality but make sure you satisfy my expectation for it to record the user so that everyone else can hear what others are chanting.
So start with Unmute Myself and only if the user wants themselves to be heard, then the Unmute Myself button will Unmute the users for others.
Ask user for microphone access as soon as click on Unmute Myself.
a person should be able to hear other people chanting as soon as they click on the Start Chanting button.
How do I test this? How do I hear other people chanting?

Also do some AI magic so that someone who is chanting out of sync will be heard as being in sync by others who join the page. Something to do this should be in the code.
How do I test this?